Hybrid prime testing algorithm.
Trial by division for small numbers. Mill-Rabin for large numbers.

+++ src/solutions/self_powers.rs
@@ -3,7 +3,7 @@ use crate::utils;
 pub fn solve() -> i64 {
     let modulus = 10i64.pow(10);
     let sum = (1..1000).filter(|num| num % 10 != 0).fold(0, |sum, num| {
-        (sum + utils::pow(num, num as u32, modulus)) % modulus
+        (sum + utils::pow(num, num as u64, modulus)) % modulus
     });
 
     assert_eq!(sum, 9110846700);
